fix: pr remarks address autocomplete
