+++ packages/web/design-system/src/ui/address-autocomplete/index.ts
@@ -10,4 +10,3 @@ export {
   formattedAddressToString,
 } from './addressAutocomplete.util'
 export { default as UIAddressAutocomplete } from './AddressAutocomplete.vue'
-export { default as UIAddressAutocompleteProvider } from './AddressAutocompleteProvider.vue'
