+++ packages/web/design-system/src/ui/config-provider/ConfigProvider.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { toComputedRefs } from '@/composables/context.composable'
+import { useProvideAddressAutocompleteAdapter } from '@/ui/address-autocomplete/addressAutocomplete.context'
 import { useProvideConfigContext } from '@/ui/config-provider/config.context'
 import type { ConfigProviderProps } from '@/ui/config-provider/config.types'
 import TooltipProvider from '@/ui/tooltip/TooltipProvider.vue'
@@ -14,6 +15,10 @@ defineSlots<{
 }>()
 
 useProvideConfigContext(toComputedRefs(props))
+
+if (props.addressAutocompleteAdapter != null) {
+  useProvideAddressAutocompleteAdapter(props.addressAutocompleteAdapter)
+}
 </script>
 
 <template>
